+++ CHANGELOG.md
@@ -38,9 +38,10 @@
   there being no corresponding `ComputableValueActivatedEvent` spy event. This even is no longer generated.
 * **\[core\]** Improve invariant failure messages when attempting to add or remove `OnDispose` listeners
   to the `DisposeNotifier`. This is typically exhibited when using the annotation processor to process a class
-  that has a `@ComponentDependency` that is passed to the constructor and the dependency is already disposed. 
+  that has a `@ComponentDependency` that is passed to the constructor and the dependency is already disposed.
 * **\[core\]** Rename the annotation `@OnDepsChanged` to `@OnDepsChange` to keep consistent tense in
   annotation naming convention.
+* **\[core\]** Add the `throwable` parameter to the `ObserveCompletedEvent` spy event.
 
 ### [v0.111](https://github.com/arez/arez/tree/v0.111) (2018-10-19)
 [Full Changelog](https://github.com/arez/arez/compare/v0.110...v0.111)